feat: Refactor HomePage + useTasks hook to handle functionality
